Moved all controls under Profiles
