+++ surface/src/components/Profiles/Profiles.tsx
@@ -1,15 +1,17 @@
 import * as React from 'react';
 import './Toggle.scss';
 import test from './configs/test.json';
-import Toggle from '../Toggle/Toggle';
+import ThrustTweaker from './Tweaker/ThrustTweaker';
+import Compensator from './Compensator/Compensator';
+import CoM from './CoM/CoM';
 
 const Profiles: React.FC = () => {
 
-    const [selected, setSelected] = React.useState([false])
-
     return(
         <div>
-            <Toggle name={test.name} callback={() => selected[0] = !selected[0]} toggled={selected[0]}/>
+            <ThrustTweaker/>
+            <Compensator/>
+            <CoM/>
         </div>
     )
 }
